Move benchmark code into cli commands file
