+++ memflow-cli/Cargo.toml
@@ -18,3 +18,5 @@ clap = { version = "2.33" }
 log = "0.4.8"
 simple_logger = "1.11"
 serde_json = "1.0"
+futures = "0.3"
+tokio = { version = "1.0", features = ["rt-multi-thread", "time", "fs", "macros", "net"] }
